BB - Updated Enumeration Migrations. Added logging and Delete Enumeration value
